fix(recon): resolve the 13 Medium-severity code-review findings
Real fixes (grounded in fresh disasm, verified with a build):
- sprite_anim.c: sprite_expand_frame's Path B read width/height as single
  bytes; the real asm (1cec:2ee9/2ef5) reads full 16-bit words. Retyped u16
  (harmless for real BUMSPJEU data, but now faithful).
- game.c/level.c: current_level_index was typed u8; three independent asm
  sites (init, game_loop, load_current_level_data) all MOV AH,0 before a
  full-word store/load of DGROUP 0x7310 — it's a genuine u16 field. Retyped
  throughout, dropped the now-inaccurate (u8) cast at the assignment site.
- game_stubs.c/input.c: dos_abort fabricated a `return 0`. The real function
  (1000:762a) is INT 21h AH=9 (print string) then JMP $ — an infinite
  noreturn halt. Retyped void, fixed both call sites.
- op12.c: op4_handler's recon-invented snapshot buffer (g_op4_scratch, no
  engine counterpart — the reference op12_port.py uses an unbounded Python
  slice) had no bound check against a data-derived length. Added a clamp;
  inert for real game data, no longer silently unsafe.

Documented (verified real but inert, or verified already-correct, no
behavior change):
- midi.c: opl_write_reg's real PUSHF;CLI...POPF;RET bracket, hidden by a
  Ghidra function-boundary bug that cuts off mid-body; opl_play_note's DH
  ambient-register assumption, traced back through its one caller with no
  oracle to confirm it; midi_read_varlen's 3-byte VLQ branch has the same
  deterministic off-by-(-0x40) bug as the already-documented 4-byte branch
  (triggers iff bit 6 of the third byte is set) — corrected an earlier note
  that overclaimed the 3-byte branch was clean.
- screens.c: read_chunked's call sites push a 6th always-zero word the
  callee never reads — a provably inert original-binary vestige, documented
  at the shared prototype. The menu/in-game sprite-buffer literal-address
  writes (vs. the real asm's memory reads) are confirmed deliberate and
  load-bearing per host_render.c's cursor/front-end routing, not bugs.
- gfx_overlay.c, level.c's load_current_level_data buffer placement, and
  game_stubs.c's timing-function carve-out ownership comment: verified
  correct or already fully documented; fixed one stale comment (host_timer.c
  -> game.c, after the relocation).

Refuted: op12.c's vec_read_record dispatch has no missing table-bounds
check — no dispatch table exists at DGROUP 0x4e5b or anywhere else, in
either the original engine or the already-validated op12_port.py reference;
dispatch is a plain two-way if/else if, exactly mirrored in op12.c.

All touched files compile clean under both build flag sets; wmake all and
wmake play both link successfully.

+++ src/game.c
@@ -73,7 +73,13 @@ u8 session_continue_flag;    /* DGROUP 0x856d (DAT_856d) */
 u8 frame_abort_flag;         /* DGROUP 0x928d (DAT_203b_928d) */
 
 /* ── game_loop / init game-state globals OWNED here (no single module home) ──── */
-u8  current_level_index;     /* DGROUP 0x7310 — 0-based level index (= current_level-1) */
+/* RECONSTRUCTION FIDELITY (verified 2026-07-14): DGROUP 0x7310 is a genuine 16-bit
+ * field — every asm access (1000:030b/030d init-zero, 1000:0c72..0c78 game_loop,
+ * 1000:03b1/03cb + 1000:32bc/32d6 load_current_level_data) loads/stores the full AX,
+ * with an explicit `MOV AH,0` zero-extend from the byte source, never a lone byte
+ * MOV.  Typed u16 here (not u8) for signature fidelity; harmless in practice since
+ * every real level index is 0-8. */
+u16 current_level_index;     /* DGROUP 0x7310 — 0-based level index (= current_level-1) */
 u8  palette_loaded;          /* DGROUP — set 0 by init, by apply_level_palette later   */
 u8  menu_option2_setting;    /* game_loop menu scratch */
 u8  settle_countdown;        /* DGROUP 0x791a — ONE engine byte, four users (unified
@@ -460,7 +466,7 @@ LAB_0c2c:
                         show_text_screen();
                         goto LAB_0c2c;
                     }
-                    current_level_index = (u8)(current_entity_index - 1);
+                    current_level_index = (u16)(current_entity_index - 1);
                     play_iris_wipe_transition();
                     reset_game_state();
                     p2_set_move_state(p2_move_state);
